helium: add libion dependancy
Default public build for Helium board does not generate proprietary
Marvell graphics binaries.  During this build libion is never generated
as a result.

Let's add a libion dependancy, so that it will be present in system
if the user flashes a vendor.img containing HW graphic support libs.

This fixes a Marvell gralloc module load issue:
dlopen failed: library "libion.so" not found

Signed-off-by: Michael Scott <michael.scott@linaro.org>